Create unique tags if tag already exists.
When two repositories have the same tag pointing to different
commits, the script creates a merge commit and moves the tags
but it creates two tags to identify the old commits.

If those tags exist the script previously failed. However using
the script to merge more than one repository into a specific
repository will probably lead to that state that when merging
the second repository the 'unique' tags might already exist.

Now the script ammends an integer (using the first available)
to be able to create a tag anyway.